feat: move ideas files from command-center into the new ideas app (closes #289)
